Refactor Scaffoling again
