--- genomix/genomix-pregelix/src/main/java/edu/uci/ics/genomix/pregelix/io/common/ByteBacked.java
@@ -1,6 +1,0 @@
-package edu.uci.ics.genomix.pregelix.io.common;
-
-public abstract interface ByteBacked<E> {
-    public byte getByte();
-    public E fromByte(byte b);
-}
